--- a/cli/beamable.common/Runtime/Dependencies/Utils.cs
+++ b/cli/beamable.common/Runtime/Dependencies/Utils.cs
@@ -38,10 +38,12 @@ namespace Beamable.Common.Dependencies
 	{
 		public Type Interface, Implementation;
 		public Func<IDependencyProvider, object> Factory;
+		public DependencyLifetime Lifetime;
+		
 
 		public ServiceDescriptor Clone()
 		{
-			return new ServiceDescriptor { Interface = Interface, Implementation = Implementation, Factory = Factory };
+			return new ServiceDescriptor { Interface = Interface, Implementation = Implementation, Factory = Factory, Lifetime = Lifetime };
 		}
 	}
 }